move LOG OUT button to MenuSide
